Removed some useless classes. Made the constructor for the RubiksCube. Made the RubiksPiece class. Added the setColors() method to the FacePanel.
The other piece classes could all be made as the same class, so they were removed. Added a getter to the controller to retrieve the RubiksCube's 3d array. Set all of the values of the 3d array. Added what may be the only necessary code to the RubiksPiece. Added a constant list of colors to the FacePanel class. Added the setColors() method to the FacePanel class. This method will set the colors of the FacePanels to the corresponding spots on the RubiksCube's 3d array. Added a getter to the RubiksButton for use in the setColors() method.