Started implementing the RotateLayersPanel.
Added some base code to the RotateCubePanel. Added methods to the RubiksController to help call other methods. Added two more numbers to every button's id to indicate which layers would be turned if it was selected. Started adding the look of the RotateLayersPanel. Removed the abstract RotatePanel class. Added the selected boolean to the RubiksButton class and added handling for it. Added a getter to the RubiksFrame for the RubiksPanel. Added the deselect() and findSelected() methods to the RubiksPanel, both of which can be called by the RubiksController.