Fixed some problems with the CubeInfoPanel and added a victory condition.
Modified the scramble method so it scrambled a bit better. Added a call to the RubiksController to detect a victory. Modified the timer actionPerformed() to actually work when it limits the time. Made the startGame method set the time to 0 properly. Added a limit to the moveCount. Added the beginnings of a victory method, still needs work. Added methods to the CubePanel and FacePanel classes to detect if a victory actually occured.